requirements.txt
# Librerías necesarias para el proyecto de reconocimiento facial
numpy
Pillow
opencv-python
opencv-contrib-python